+++ Rdmp.Core/Providers/CatalogueProblemProvider.cs
@@ -121,6 +121,11 @@ namespace Rdmp.Core.Providers
                 if (!_usedJoinables.Contains(aggregateConfiguration.JoinableCohortAggregateConfiguration.ID))
                     return "Patient Index Table is not joined to any cohort sets";
 
+            if(!aggregateConfiguration.AggregateDimensions.Any())
+            {
+                return "Aggregate has no dimensions.  Set an AggregateDimension to specify which column is fetched by the query.";
+            }
+
             return null;
         }